🐛 fix: 컴포넌트 분리 수정
