+++ src/components/ui/common/Menu/index.tsx
@@ -1,21 +1,15 @@
 'use client';
 import { useEffect, useRef } from 'react';
-import { useRouter } from 'next/navigation';
-import Image from 'next/image';
-
 import { cn } from '@/lib/utils';
-import CloseIcon from '@/assets/icons/common/close.svg';
 import { useMenuContext } from '@/context/MenuContext';
-import { Body, Caption, SubTitle } from '../Typography';
-import { mainCategoryDatas } from '@/data/main/initData';
-import MoreButton from '../../main/MoreButton';
+import MenuText from '@/components/modules/Menu/MenuText';
+import MenuNav from '@/components/modules/Menu/MenuCategory';
+import MenuList from '@/components/modules/Menu/MenuList';
+import MenuHeader from '@/components/layouts/Menu/MenuHeader';
 
 export default function Menu() {
-  const router = useRouter();
   const asideRef = useRef<HTMLElement>(null);
   const { isOpen, setIsOpen } = useMenuContext();
-
-  const onClick = () => setIsOpen((prev) => !prev);
 
   useEffect(() => {
     const container = document.getElementById('container');
@@ -28,11 +22,6 @@ export default function Menu() {
       }
     };
   }, [isOpen]);
-
-  const handleRouteChange = (href: string) => {
-    setIsOpen(false);
-    router.push(href);
-  };
 
   useEffect(() => {
     const handleClickOutside = (e: MouseEvent) => {
@@ -50,21 +39,6 @@ export default function Menu() {
       document.removeEventListener('mousedown', handleClickOutside);
     };
   }, [isOpen, setIsOpen]);
-
-  const menuListData = [
-    {
-      id: 1,
-      href: '/events',
-      title: '기획전',
-      desc: '진행중인 기획전을 만나보세요.',
-    },
-    {
-      id: 2,
-      href: '/best',
-      title: '베스트',
-      desc: '스타벅스 베스트 MD 상품만 모아보세요.',
-    },
-  ];
 
   return (
     <div
@@ -86,63 +60,14 @@ export default function Menu() {
           isOpen ? 'translate-x-0 shadow-1' : '-translate-x-full shadow-none',
         )}
       >
-        <header className='flex justify-end px-6 py-4'>
-          <button onClick={onClick} className='cursor-pointer'>
-            <CloseIcon width={24} height={24} fill='var(--color-text-900)' />
-          </button>
-        </header>
+        <MenuHeader />
 
-        <section className='px-6 py-5 space-y-3'>
-          <SubTitle>Welcome !</SubTitle>
-          <Body level={3}>온라인 스토어에 오신 것을 환영합니다.</Body>
-        </section>
-
+        <MenuText title='Welcome !' text='온라인 스토어에 오신 것을 환영합니다.' />
         <hr className='mx-6' />
-
-        <nav className='px-6 py-8'>
-          <div className='grid justify-end pb-4 text-text-900 hover:text-black focus:text-black transition-colors'>
-            <button onClick={() => setIsOpen(false)}>
-              <MoreButton href='/products' title='전체 상품 보기' />
-            </button>
-          </div>
-
-          <ul className='grid grid-cols-3 md:grid-cols-4 justify-items-center items-start sm:items-center gap-x-2 gap-y-5'>
-            {mainCategoryDatas.map(({ id, title, icon }) => (
-              <li key={id} className='inline-block w-full max-w-[100px]'>
-                <button
-                  onClick={() => handleRouteChange(`/products?topCategoryId=${id}`)}
-                  className='space-y-2.5 cursor-pointer w-full'
-                >
-                  <div className='w-full aspect-square relative'>
-                    <Image
-                      src={icon}
-                      alt={`${title} 카테고리 썸네일 이미지`}
-                      sizes='100%'
-                      fill
-                      className='rounded-full object-cover'
-                    />
-                  </div>
-                  <Body level={3} className='text-center'>
-                    {title}
-                  </Body>
-                </button>
-              </li>
-            ))}
-          </ul>
-        </nav>
-
-        <section className='px-6 bg-gray-200'>
-          <ul className='pb-16'>
-            {menuListData.map(({ id, href, title, desc }) => (
-              <li key={id}>
-                <button onClick={() => handleRouteChange(href)}>
-                  <Caption level={1}>{title}</Caption>
-                  <Caption level={3}>{desc}</Caption>
-                </button>
-              </li>
-            ))}
-          </ul>
-        </section>
+        <div className='grid  grid-rows-2'>
+          <MenuNav />
+          <MenuList />
+        </div>
       </aside>
     </div>
   );
